formats: move Log* to a separate package
